Btrfs: send, truncate to shrink before writing when doing incremental send
Currently, when a file changes size, we send a truncate in
finish_inode_in_needed(). However, an upcoming change needs to move the
truncate elsewhere. For the case where the file grows, this requires a
few extra steps. But, when the file shrinks, we can send the truncate
before doing anything else.

This also ensures that when we add an fallocate command, we send the
truncate before we create any preallocated extents beyond i_size
(otherwise, the truncate would remove the preallocated extents).

Signed-off-by: Omar Sandoval <osandov@fb.com>