Btrfs: send, send a hole for remaining extents beyond EOF
In finish_inode_if_needed(), we send a hole for the rest of the file
that we haven't handled yet, up to i_size. However, if we want to
preserve the layout of preallocated extents beyond i_size, we also need
to punch a hole in any preallocated extents beyond EOF that were
deleted. So, instead of sending a hole up to i_size, send it up to the
maximum offset; it will be trimmed down by non_hole_range_in_parent().

Currently, this is a no-op because we don't write zeroes beyond EOF.
However, once we support the fallocate command, this will punch holes
when necessary.

Signed-off-by: Omar Sandoval <osandov@fb.com>